Rabin_Karp

+++ KMP/KMP.c
@@ -11,7 +11,7 @@
 void compute_prefix(const char* pattern, int next[]) {
 	int i;
 	int j = -1;
-	const int m = strlen(pattern);
+	const int m = (int)strlen(pattern);
 	next[0] = j;
 	for (i = 1; i < m; i++) {
 		while (j > -1 && pattern[j + 1] != pattern[i]) j = next[j];
@@ -30,8 +30,8 @@ int kmp(const char* text, const char* pattern)
 {
 	int i;
 	int j = -1;
-	const int n = strlen(text);
-	const int m = strlen(pattern);
+	const int n = (int)strlen(text);
+	const int m = (int)strlen(pattern);
 	if (n == 0 && m == 0) return 0; /* "","" */
 	if (m == 0) return 0; /* "a","" */
 	int* next = (int*)malloc(sizeof(int) * m);
